Set GOPATH with godeps on standard run
This change sets GOPATH to use the godep path by default. The godeps target
simply sets the GOPATH then runs all the targets. This allows one to get
reliable release builds by default while at the same time allowing external
manipulations of the GOPATH to be used with the Makefile.

Signed-off-by: Stephen J Day <stephen.day@docker.com>